docker client added

+++ build_image.sh
@@ -5,20 +5,28 @@ GREEN='\033[0;32m'
 RED='\033[0;31m'
 NC='\033[0m'
 
+NAMESPACE="elegoev"
 CONTAINERNAME="jenkinsci"
 
 # create image
 echo -e "${GREEN}#########################${NC}"
 echo -e "${GREEN}# >>>> remove container #${NC}"
 echo -e "${GREEN}#########################${NC}"
-docker stop $CONTAINERNAME
-docker rm $CONTAINERNAME
+CONTAINERID=$(docker ps -a | grep $CONTAINERNAME | awk '{ print $1 }')
+if [ -z $CONTAINERID ]
+then
+  echo -e "Container $CONTAINERNAME doesn't exists"
+else
+  docker stop $CONTAINERNAME
+  docker rm $CONTAINERNAME
+  echo -e "Container $CONTAINERNAME deleted"
+fi
 
 # create image
 echo -e "${GREEN}######################${NC}"
 echo -e "${GREEN}# >>>> create image  #${NC}"
 echo -e "${GREEN}######################${NC}"
-docker build --rm -t elegoev/jenkinsci .
+docker build --rm -t $NAMESPACE/$CONTAINERNAME .
 
 # run image
 echo -e "${GREEN}######################${NC}"
@@ -26,4 +34,5 @@ echo -e "${GREEN}# >>>> run image     #${NC}"
 echo -e "${GREEN}######################${NC}"
 docker run -d --restart=unless-stopped --name $CONTAINERNAME \
            -p 8080:8080 -p 50000:50000 \
-           elegoev/jenkinsci
+           -v /var/run/docker.sock:/var/run/docker.sock \
+           $NAMESPACE/$CONTAINERNAME
